+++ src/main/java/com/comeon/backend/image/presentation/request/ImageRemoveRequest.java
@@ -1,4 +1,4 @@
-package com.comeon.backend.image.presentation.api.request;
+package com.comeon.backend.image.presentation.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
