+++ src/main/java/com/comeon/backend/image/presentation/response/ImageRemoveResponse.java
@@ -1,4 +1,4 @@
-package com.comeon.backend.image.presentation.api.response;
+package com.comeon.backend.image.presentation.response;
 
 import lombok.Getter;
